isl_multi_templ.c: extract out isl_multi_product_templ.c
This makes the function isl_multi_*_product opt-in rather than opt-out.
