+++ isl_multi_templ.c
@@ -857,69 +857,6 @@ __isl_give MULTI(BASE) *FN(MULTI(BASE),factor_range)(
 	return multi;
 }
 
-#ifndef NO_PRODUCT
-/* Given two MULTI(BASE)s A -> B and C -> D,
- * construct a MULTI(BASE) [A -> C] -> [B -> D].
- *
- * The parameters are assumed to have been aligned.
- *
- * If "multi1" and/or "multi2" has an explicit domain, then
- * intersect the domain of the result with these explicit domains.
- */
-__isl_give MULTI(BASE) *FN(MULTI(BASE),product_aligned)(
-	__isl_take MULTI(BASE) *multi1, __isl_take MULTI(BASE) *multi2)
-{
-	int i;
-	EL *el;
-	isl_space *space;
-	MULTI(BASE) *res;
-	int in1, in2, out1, out2;
-
-	in1 = FN(MULTI(BASE),dim)(multi1, isl_dim_in);
-	in2 = FN(MULTI(BASE),dim)(multi2, isl_dim_in);
-	out1 = FN(MULTI(BASE),dim)(multi1, isl_dim_out);
-	out2 = FN(MULTI(BASE),dim)(multi2, isl_dim_out);
-	space = isl_space_product(FN(MULTI(BASE),get_space)(multi1),
-				  FN(MULTI(BASE),get_space)(multi2));
-	res = FN(MULTI(BASE),alloc)(isl_space_copy(space));
-	space = isl_space_domain(space);
-
-	for (i = 0; i < out1; ++i) {
-		el = FN(FN(MULTI(BASE),get),BASE)(multi1, i);
-		el = FN(EL,insert_dims)(el, isl_dim_in, in1, in2);
-		el = FN(EL,reset_domain_space)(el, isl_space_copy(space));
-		res = FN(FN(MULTI(BASE),set),BASE)(res, i, el);
-	}
-
-	for (i = 0; i < out2; ++i) {
-		el = FN(FN(MULTI(BASE),get),BASE)(multi2, i);
-		el = FN(EL,insert_dims)(el, isl_dim_in, 0, in1);
-		el = FN(EL,reset_domain_space)(el, isl_space_copy(space));
-		res = FN(FN(MULTI(BASE),set),BASE)(res, out1 + i, el);
-	}
-
-	if (FN(MULTI(BASE),has_explicit_domain)(multi1) ||
-	    FN(MULTI(BASE),has_explicit_domain)(multi2))
-		res = FN(MULTI(BASE),intersect_explicit_domain_product)(res,
-								multi1, multi2);
-
-	isl_space_free(space);
-	FN(MULTI(BASE),free)(multi1);
-	FN(MULTI(BASE),free)(multi2);
-	return res;
-}
-
-/* Given two MULTI(BASE)s A -> B and C -> D,
- * construct a MULTI(BASE) [A -> C] -> [B -> D].
- */
-__isl_give MULTI(BASE) *FN(MULTI(BASE),product)(
-	__isl_take MULTI(BASE) *multi1, __isl_take MULTI(BASE) *multi2)
-{
-	return FN(MULTI(BASE),align_params_multi_multi_and)(multi1, multi2,
-					&FN(MULTI(BASE),product_aligned));
-}
-#endif
-
 __isl_give MULTI(BASE) *FN(MULTI(BASE),flatten_range)(
 	__isl_take MULTI(BASE) *multi)
 {
